requirements.txt
decorator==4.3.2
dicttoxml==1.7.4
Jinja2==2.10
MarkupSafe==1.1.0
PyYAML==3.13
six==1.12.0
validators==0.12.4
